payment/requirements.txt
Flask==3.0.2
psycopg2-binary
gunicorn==21.2.0
msgspec==0.18.6